requirements.txt
sqlalchemy
openai
slack-sdk
python-dotenv
schedule
flask